+++ src/main/kotlin/app/revanced/patches/reddit/customclients/boostforreddit/fix/slink/FixSLinksPatch.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.reddit.customclients.syncforreddit.fix.slink
+package app.revanced.patches.reddit.customclients.boostforreddit.fix.slink
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
@@ -7,28 +7,23 @@ import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.fingerprint.MethodFingerprintResult
 import app.revanced.patcher.util.smali.ExternalLabel
 import app.revanced.patches.reddit.customclients.BaseFixSLinksPatch
-import app.revanced.patches.reddit.customclients.syncforreddit.fix.slink.fingerprints.LinkHelperOpenLinkFingerprint
-import app.revanced.patches.reddit.customclients.syncforreddit.fix.slink.fingerprints.SetAuthorizationHeaderFingerprint
-import app.revanced.patches.reddit.customclients.syncforreddit.misc.integrations.IntegrationsPatch
+import app.revanced.patches.reddit.customclients.boostforreddit.fix.slink.fingerprints.GetOAuthAccessTokenFingerprint
+import app.revanced.patches.reddit.customclients.boostforreddit.fix.slink.fingerprints.HandleNavigationFingerprint
+import app.revanced.patches.reddit.customclients.boostforreddit.misc.integrations.IntegrationsPatch
 
 @Suppress("unused")
 object FixSLinksPatch : BaseFixSLinksPatch(
-    handleNavigationFingerprint = LinkHelperOpenLinkFingerprint,
-    setAccessTokenFingerprint = SetAuthorizationHeaderFingerprint,
-    compatiblePackages = setOf(
-        CompatiblePackage("com.laurencedawson.reddit_sync"),
-        CompatiblePackage("com.laurencedawson.reddit_sync.pro"),
-        CompatiblePackage("com.laurencedawson.reddit_sync.dev"),
-    ),
+    handleNavigationFingerprint = HandleNavigationFingerprint,
+    setAccessTokenFingerprint = GetOAuthAccessTokenFingerprint,
+    compatiblePackages = setOf(CompatiblePackage("com.rubenmayayo.reddit")),
     dependencies = setOf(IntegrationsPatch::class),
 ) {
-    override val integrationsClassDescriptor = "Lapp/revanced/integrations/syncforreddit/FixSLinksPatch;"
+    override val integrationsClassDescriptor = "Lapp/revanced/integrations/boostforreddit/FixSLinksPatch;"
 
     override fun MethodFingerprintResult.patchNavigationHandler(context: BytecodeContext) {
         mutableMethod.apply {
-            val urlRegister = "p3"
-            val tempRegister = "v2"
-
+            val urlRegister = "p1"
+            val tempRegister = "v1"
             addInstructionsWithLabels(
                 0,
                 """
@@ -43,7 +38,7 @@ object FixSLinksPatch : BaseFixSLinksPatch(
     }
 
     override fun MethodFingerprintResult.patchSetAccessToken(context: BytecodeContext) = mutableMethod.addInstruction(
-        0,
-        "invoke-static { p0 }, $integrationsClassDescriptor->$setAccessTokenMethod",
+        3,
+        "invoke-static { v0 }, $integrationsClassDescriptor->$setAccessTokenMethod",
     )
 }